[HER-15]Follow-up Add different style files
