+++ src/user/row/style.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core'
+
+const paper = {
+  paddingTop: '1vh',
+  paddingBottom: '1vh',
+  textAlign: 'center',
+  fontSize: '1vw',
+}
+
+export const useStyle = makeStyles(() => ({ paper }))
