+++ src/user/style.js
@@ -1,16 +1,8 @@
 import { makeStyles } from '@material-ui/core'
 
-const title = {
-  textAlign: 'center',
-  fontSize: '2vw',
+const main = {
+  maxWidth: '80%',
+  margin: 'auto',
 }
 
-const paper = {
-  paddingTop: '1vh',
-  paddingBottom: '1vh',
-  textAlign: 'center',
-  fontSize: '1vw',
-}
-
-const main = { maxWidth: '80%', margin: 'auto' }
-export const useStyle = makeStyles(() => ({ title, paper, main }))
+export const useStyle = makeStyles(() => ({ main }))
